pyproject.toml
[project]
name = "work-morables"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "datasets",
    "sentence-transformers",
    "matplotlib",
    "pandas",
    "numpy",
    "scikit-learn",
    "tqdm",
    "wandb",
    "pyyaml",
    "peft",
    "google-generativeai",
    "InstructorEmbedding",
    "einops",
    "torch",
    "transformers",
    "gritlm",
    "mlx-lm ; sys_platform == 'darwin'",
]

[tool.pytest.ini_options]
testpaths = ["tests", "experiments"]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[tool.uv.sources]
mlx-lm = { git = "https://github.com/ml-explore/mlx-lm.git" }
torch = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux'" },
]
